requirements.txt
feedparser
feedgen
listparser
pyyaml
yt-dlp
matplotlib
numpy
